original selection code separator. Only used for xPhoton code. (Prevent big change in code
